+++ src/main/java/Controller/Controller.java
@@ -1,0 +1,9 @@
+package Controller;
+
+import http.request.HttpRequest;
+import http.response.HttpResponse;
+
+public interface Controller {
+
+    String process(HttpRequest request, HttpResponse httpResponse);
+}
