--- src/main/java/handler/Handler.java
@@ -1,9 +1,0 @@
-package handler;
-
-import http.request.HttpRequest;
-import http.response.HttpResponse;
-
-public interface Handler {
-
-    String handle(HttpRequest request, HttpResponse httpResponse);
-}
